Logger: Separate forwarder code
