+++ libs/qec/lib/logger.cpp
@@ -7,29 +7,27 @@
  ******************************************************************************/
 
 #include "cudaq/qec/logger.h"
+#include "logger_forwarder.h"
 #include <algorithm>
-#include <array>
 #include <atomic>
 #include <cctype>
 #include <chrono>
-#include <condition_variable>
 #include <cstdint>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
 #include <ctime>
 #include <iomanip>
-#include <memory>
 #include <mutex>
 #include <optional>
 #include <sstream>
 
 /// @file
-/// @brief QEC logger implementation.
+/// @brief Core QEC logger front-end.
 /// @details
-/// Implements runtime level filtering, synchronous stdout/stderr emission,
-/// optional asynchronous forwarding to a worker thread, and helper utilities
-/// for formatting timestamped log lines with source metadata.
+/// Owns level parsing, timestamp/source formatting, and synchronous sink
+/// output. When forwarding is enabled, this file packs records into bounded
+/// buffers and delegates asynchronous delivery to `logger_forwarder.cpp`.
 
 namespace cudaq::qec::detail {
 namespace {
@@ -39,14 +37,14 @@ using Clock = std::chrono::system_clock;
 std::atomic<LogLevel> gLogLevel{LogLevel::warn};
 std::once_flag gLogLevelInitFlag;
 
-// Convert a level token to lowercase before parsing.
+// Convert a log level token to lowercase for robust parsing.
 std::string toLower(std::string value) {
   for (auto &ch : value)
     ch = static_cast<char>(std::tolower(static_cast<unsigned char>(ch)));
   return value;
 }
 
-// Parse text level from env/config to an internal enum.
+// Parse CUDAQ_LOG_LEVEL text into internal enum.
 std::optional<LogLevel> parseLogLevel(const std::string_view level) {
   const std::string lower = toLower(std::string(level));
   if (lower == "trace")
@@ -60,9 +58,9 @@ std::optional<LogLevel> parseLogLevel(const std::string_view level) {
   if (lower == "error")
     return LogLevel::error;
   return std::nullopt;
-} // end - parseLogLevel()
+}
 
-// Convert internal log level to stable label in log lines.
+// Convert internal level enum to stable output label.
 const char *logLevelName(const LogLevel level) {
   switch (level) {
   case LogLevel::trace:
@@ -77,9 +75,9 @@ const char *logLevelName(const LogLevel level) {
     return "error";
   }
   return "info";
-} // end - logLevelName()
+}
 
-// Return current wall-clock timestamp in nanoseconds since epoch.
+// Return current wall-clock timestamp in nanoseconds.
 std::uint64_t nowNs() {
   return static_cast<std::uint64_t>(
       std::chrono::duration_cast<std::chrono::nanoseconds>(
@@ -87,7 +85,7 @@ std::uint64_t nowNs() {
           .count());
 }
 
-// Render local-time timestamp with microsecond precision for log prefixes.
+// Render local-time timestamp with microsecond precision.
 std::string formatTimestamp(const Clock::time_point tp) {
   const auto seconds = std::chrono::time_point_cast<std::chrono::seconds>(tp);
   const auto micros = std::chrono::duration_cast<std::chrono::microseconds>(
@@ -105,9 +103,9 @@ std::string formatTimestamp(const Clock::time_point tp) {
   out << std::put_time(&tm, "%Y-%m-%d %H:%M:%S") << '.' << std::setw(6)
       << std::setfill('0') << micros;
   return out.str();
-} // end - formatTimestamp()
+}
 
-// Build one final log line with timestamp, level, source, and payload.
+// Build one final text log line with timestamp/source metadata.
 std::string composeLine(const LogLevel level, const std::string_view message,
                         const char *fileName, const int lineNo,
                         const Clock::time_point ts) {
@@ -117,6 +115,7 @@ std::string composeLine(const LogLevel level, const std::string_view message,
   return out.str();
 }
 
+// Lightweight filename extraction view to avoid allocations.
 std::string_view pathToFileNameView(const std::string_view fullFilePath) {
   const auto pos = fullFilePath.find_last_of("/\\");
   if (pos == std::string_view::npos)
@@ -138,305 +137,7 @@ std::size_t copyToFixed(std::array<char, N> &dest, const std::string_view src,
   return copied;
 }
 
-struct QueuedLogRecord {
-  LogLevel level = LogLevel::info;
-  std::uint64_t timestampNs = 0;
-  int lineNo = 0;
-  std::thread::id threadId;
-  std::array<char, 128> fileName{};
-  std::size_t fileNameLen = 0;
-  std::array<char, detail::kRealtimeForwarderMaxMessageCapacity> message{};
-  std::size_t messageLen = 0;
-};
-
-class AsyncForwarder {
-public:
-  // Default-construct the forwarder with forwarding disabled.
-  AsyncForwarder() = default;
-  // Ensure worker thread is stopped before object destruction.
-  ~AsyncForwarder() { clear(); }
-
-  // Install callback + queue configuration and start worker thread.
-  void set(ForwarderConfig config) {
-    clear();
-    mConfig = std::move(config);
-    if (!mConfig.callback)
-      return;
-    if (mConfig.queueCapacity == 0)
-      mConfig.queueCapacity = 1;
-    if (mConfig.messageCapacity == 0)
-      mConfig.messageCapacity = 1;
-    mConfig.messageCapacity =
-        std::min(mConfig.messageCapacity, kRealtimeForwarderMaxMessageCapacity);
-    mMessageCapacity.store(mConfig.messageCapacity, std::memory_order_relaxed);
-
-    mCapacity = roundUpToPow2(mConfig.queueCapacity);
-    mRing = std::make_unique<RingSlot[]>(mCapacity);
-    for (std::size_t i = 0; i < mCapacity; ++i)
-      mRing[i].sequence.store(i, std::memory_order_relaxed);
-
-    mMask = mCapacity - 1;
-    mEnqueuePos.store(0, std::memory_order_relaxed);
-    mDequeuePos.store(0, std::memory_order_relaxed);
-    mStop.store(false, std::memory_order_relaxed);
-    mEnabled.store(true, std::memory_order_release);
-    mWorker = std::thread([this] { runWorker(); });
-  } // end - set()
-
-  // Stop worker thread and reset forwarding state.
-  void clear() {
-    mEnabled.store(false, std::memory_order_release);
-    while (mActiveProducers.load(std::memory_order_acquire) != 0)
-      std::this_thread::yield();
-    mStop.store(true, std::memory_order_release);
-    mCv.notify_all();
-    if (mWorker.joinable())
-      mWorker.join();
-    mRing.reset();
-    mCapacity = 0;
-    mMask = 0;
-    mEnqueuePos.store(0, std::memory_order_relaxed);
-    mDequeuePos.store(0, std::memory_order_relaxed);
-    mStop.store(false, std::memory_order_relaxed);
-    mMessageCapacity.store(kRealtimeForwarderDefaultMessageCapacity,
-                           std::memory_order_relaxed);
-    mConfig = {};
-  } // end - clear()
-
-  // Fast path atomic check used by producers.
-  bool isEnabled() const { return mEnabled.load(std::memory_order_relaxed); }
-
-  // Return lock-free counters snapshot for diagnostics/tests.
-  ForwarderStats stats() const {
-    ForwarderStats stats;
-    stats.enqueuedRecords = mEnqueuedRecords.load(std::memory_order_relaxed);
-    stats.droppedRecords = mDroppedRecords.load(std::memory_order_relaxed);
-    stats.truncatedRecords = mTruncatedRecords.load(std::memory_order_relaxed);
-    stats.forwardFailures = mForwardFailures.load(std::memory_order_relaxed);
-    return stats;
-  }
-
-  // Reset counters when installing a new forwarder config.
-  void resetStats() {
-    mEnqueuedRecords.store(0, std::memory_order_relaxed);
-    mDroppedRecords.store(0, std::memory_order_relaxed);
-    mTruncatedRecords.store(0, std::memory_order_relaxed);
-    mForwardFailures.store(0, std::memory_order_relaxed);
-    mDropWarningPending.store(false, std::memory_order_relaxed);
-    mDropWarningEmitted.store(false, std::memory_order_relaxed);
-  }
-
-  std::size_t messageCapacity() const {
-    return mMessageCapacity.load(std::memory_order_relaxed);
-  }
-
-  void noteTruncation() {
-    mTruncatedRecords.fetch_add(1, std::memory_order_relaxed);
-  }
-
-  // Wait until queue drains, used in tests and explicit flush.
-  void flush() {
-    std::unique_lock<std::mutex> lock(mWaitMutex);
-    mCv.wait(lock, [&] { return isQueueEmpty() || !mEnabled.load(); });
-  }
-
-  // Producer path: enqueue into ring buffer, drop only when saturated.
-  void tryEnqueue(QueuedLogRecord record) {
-    if (!isEnabled())
-      return;
-    ProducerGuard producerGuard(mActiveProducers);
-    if (!isEnabled() || !mConfig.callback || !mRing)
-      return;
-
-    while (true) {
-      if (tryEnqueueOne(record)) {
-        mEnqueuedRecords.fetch_add(1, std::memory_order_relaxed);
-        mCv.notify_one();
-        return;
-      }
-
-      if (mConfig.dropPolicy == ForwardDropPolicy::dropOldest) {
-        QueuedLogRecord ignored;
-        if (tryDequeueOne(ignored)) {
-          notifyDrop();
-          continue;
-        }
-      }
-
-      notifyDrop();
-      return;
-    }
-  } // end - tryEnqueue()
-
-private:
-  struct RingSlot {
-    std::atomic<std::size_t> sequence{0};
-    QueuedLogRecord record;
-  };
-
-  struct ProducerGuard {
-    explicit ProducerGuard(std::atomic<std::uint64_t> &counterRef)
-        : counter(counterRef) {
-      counter.fetch_add(1, std::memory_order_acq_rel);
-    }
-    ~ProducerGuard() { counter.fetch_sub(1, std::memory_order_acq_rel); }
-    std::atomic<std::uint64_t> &counter;
-  };
-
-  static std::size_t roundUpToPow2(std::size_t value) {
-    // This queue uses `index = position & (capacity - 1)` and per-slot sequence
-    // arithmetic (Vyukov-style ring). That mapping is equivalent to modulo only
-    // for power-of-two capacities; non-power-of-two sizes break slot mapping
-    // and can corrupt full/empty detection.
-    std::size_t rounded = 2;
-    while (rounded < value)
-      rounded <<= 1;
-    return rounded;
-  }
-
-  bool tryEnqueueOne(QueuedLogRecord &record) {
-    std::size_t pos = mEnqueuePos.load(std::memory_order_relaxed);
-    while (true) {
-      RingSlot &slot = mRing[pos & mMask];
-      const std::size_t sequence =
-          slot.sequence.load(std::memory_order_acquire);
-      const std::intptr_t dif = static_cast<std::intptr_t>(sequence) -
-                                static_cast<std::intptr_t>(pos);
-
-      if (dif == 0) {
-        if (mEnqueuePos.compare_exchange_weak(pos, pos + 1,
-                                              std::memory_order_relaxed))
-          break;
-        continue;
-      }
-      if (dif < 0)
-        return false;
-      pos = mEnqueuePos.load(std::memory_order_relaxed);
-    }
-
-    RingSlot &slot = mRing[pos & mMask];
-    slot.record = std::move(record);
-    slot.sequence.store(pos + 1, std::memory_order_release);
-    return true;
-  }
-
-  bool tryDequeueOne(QueuedLogRecord &record) {
-    std::size_t pos = mDequeuePos.load(std::memory_order_relaxed);
-    while (true) {
-      RingSlot &slot = mRing[pos & mMask];
-      const std::size_t sequence =
-          slot.sequence.load(std::memory_order_acquire);
-      const std::intptr_t dif = static_cast<std::intptr_t>(sequence) -
-                                static_cast<std::intptr_t>(pos + 1);
-
-      if (dif == 0) {
-        if (mDequeuePos.compare_exchange_weak(pos, pos + 1,
-                                              std::memory_order_relaxed))
-          break;
-        continue;
-      }
-      if (dif < 0)
-        return false;
-      pos = mDequeuePos.load(std::memory_order_relaxed);
-    }
-
-    RingSlot &slot = mRing[pos & mMask];
-    record = std::move(slot.record);
-    slot.sequence.store(pos + mCapacity, std::memory_order_release);
-    return true;
-  }
-
-  bool isQueueEmpty() const {
-    return mEnqueuePos.load(std::memory_order_acquire) ==
-           mDequeuePos.load(std::memory_order_acquire);
-  }
-
-  // Emit a one-time diagnostic when forwarding drops are first observed.
-  void notifyDrop() {
-    mDroppedRecords.fetch_add(1, std::memory_order_relaxed);
-    mDropWarningPending.store(true, std::memory_order_relaxed);
-    mCv.notify_one();
-  }
-
-  void emitDropWarningIfPending() {
-    if (!mDropWarningPending.exchange(false, std::memory_order_relaxed))
-      return;
-    bool expected = false;
-    if (!mDropWarningEmitted.compare_exchange_strong(expected, true,
-                                                     std::memory_order_relaxed))
-      return;
-    std::fputs("[cudaq::qec::logger] forwarder dropped log records "
-               "(queue full); increase ForwarderConfig::queueCapacity or "
-               "reduce callback latency.\n",
-               stderr);
-  }
-
-  // Consume queued records and invoke callback on background thread.
-  void runWorker() {
-    while (true) {
-      emitDropWarningIfPending();
-      QueuedLogRecord queuedRecord;
-      if (!tryDequeueOne(queuedRecord)) {
-        std::unique_lock<std::mutex> lock(mWaitMutex);
-        mCv.wait(lock, [&] {
-          return mStop.load(std::memory_order_acquire) || !isQueueEmpty();
-        });
-        if (mStop.load(std::memory_order_acquire) && isQueueEmpty())
-          return;
-        continue;
-      }
-
-      ForwardedLogRecord record;
-      record.level = queuedRecord.level;
-      record.timestampNs = queuedRecord.timestampNs;
-      record.fileName.assign(queuedRecord.fileName.data(),
-                             queuedRecord.fileNameLen);
-      record.lineNo = queuedRecord.lineNo;
-      record.message.assign(queuedRecord.message.data(),
-                            queuedRecord.messageLen);
-      record.threadId = queuedRecord.threadId;
-
-      try {
-        if (mConfig.callback)
-          mConfig.callback(record);
-      } catch (...) {
-        mForwardFailures.fetch_add(1, std::memory_order_relaxed);
-      }
-
-      emitDropWarningIfPending();
-      mCv.notify_all();
-    } // end - while(true)
-  } // end - runWorker()
-
-  std::condition_variable_any mCv;
-  mutable std::mutex mWaitMutex;
-  ForwarderConfig mConfig;
-  std::unique_ptr<RingSlot[]> mRing;
-  std::size_t mCapacity = 0;
-  std::size_t mMask = 0;
-  std::atomic<std::size_t> mEnqueuePos{0};
-  std::atomic<std::size_t> mDequeuePos{0};
-  std::atomic<bool> mStop{false};
-  std::thread mWorker;
-  std::atomic<bool> mEnabled{false};
-  std::atomic<std::uint64_t> mEnqueuedRecords{0};
-  std::atomic<std::uint64_t> mDroppedRecords{0};
-  std::atomic<std::uint64_t> mTruncatedRecords{0};
-  std::atomic<std::uint64_t> mForwardFailures{0};
-  std::atomic<std::uint64_t> mActiveProducers{0};
-  std::atomic<bool> mDropWarningPending{false};
-  std::atomic<bool> mDropWarningEmitted{false};
-  std::atomic<std::size_t> mMessageCapacity{
-      kRealtimeForwarderDefaultMessageCapacity};
-};
-
-// Return singleton forwarder shared by all logger call sites.
-AsyncForwarder &forwarder() {
-  static AsyncForwarder instance;
-  return instance;
-}
-
-// Lazily initialize default log level from CUDAQ_LOG_LEVEL env var.
+// Lazily initialize runtime log level from CUDAQ_LOG_LEVEL once.
 void initializeLogLevelFromEnv() {
   std::call_once(gLogLevelInitFlag, [] {
     if (const char *env = std::getenv("CUDAQ_LOG_LEVEL")) {
@@ -446,12 +147,13 @@ void initializeLogLevelFromEnv() {
   });
 }
 
-// Emit to forwarding sink when enabled; otherwise use stdout/stderr.
 void emit(const LogLevel level, const std::string_view rawMessage,
           const char *fileName, const int lineNo) {
-  if (forwarder().isEnabled()) {
-    QueuedLogRecord record;
-    const std::size_t messageCap = forwarder().messageCapacity();
+  // Forwarder mode uses fixed-size packing to keep producer-side allocation
+  // predictable and bounded.
+  if (forwarder_internal::isEnabled()) {
+    forwarder_internal::QueuedLogRecord record;
+    const std::size_t messageCap = forwarder_internal::messageCapacity();
     record.level = level;
     record.timestampNs = nowNs();
     record.fileNameLen =
@@ -462,7 +164,7 @@ void emit(const LogLevel level, const std::string_view rawMessage,
     record.messageLen =
         copyToFixed(record.message, rawMessage, messageCap, &truncated);
     if (truncated) {
-      forwarder().noteTruncation();
+      forwarder_internal::noteTruncation();
       if (record.messageLen >= kRealtimeTruncationSuffix.size()) {
         const std::size_t start =
             record.messageLen - kRealtimeTruncationSuffix.size();
@@ -472,9 +174,9 @@ void emit(const LogLevel level, const std::string_view rawMessage,
       }
     }
     record.threadId = std::this_thread::get_id();
-    forwarder().tryEnqueue(std::move(record));
+    forwarder_internal::enqueue(std::move(record));
     return;
-  } // end - if(forwarder().isEnabled())
+  }
 
   const auto ts = Clock::now();
   const std::string fullLine =
@@ -483,7 +185,7 @@ void emit(const LogLevel level, const std::string_view rawMessage,
       (level == LogLevel::warn || level == LogLevel::error) ? stderr : stdout;
   std::fputs(fullLine.c_str(), stream);
   std::fputc('\n', stream);
-} // end - emit()
+}
 
 } // namespace
 
@@ -494,13 +196,13 @@ bool should_log(const LogLevel logLevel) {
          static_cast<int>(gLogLevel.load(std::memory_order_relaxed));
 }
 
-// Public API: install/replace forwarding callback.
+// Public API: install/replace asynchronous forwarding callback.
 void setForwarder(ForwarderConfig config) {
-  forwarder().set(std::move(config));
-  forwarder().resetStats();
+  forwarder_internal::set(std::move(config));
+  forwarder_internal::resetStats();
 }
 
-// Public API: enable forwarding with a default stdout/stderr callback.
+// Public API: enable forwarding with default stdout/stderr callback.
 void setForwarder() {
   setForwarder(ForwarderConfig{
       .callback =
@@ -522,42 +224,36 @@ void setForwarder() {
 }
 
 // Public API: disable asynchronous forwarding.
-void clearForwarder() { forwarder().clear(); }
-
-// Public API: report whether forwarding callback is active.
-bool isForwarderEnabled() { return forwarder().isEnabled(); }
-
+void clearForwarder() { forwarder_internal::clear(); }
+// Public API: report whether forwarding is active.
+bool isForwarderEnabled() { return forwarder_internal::isEnabled(); }
 std::size_t getForwarderMessageCapacity() {
-  return forwarder().messageCapacity();
+  return forwarder_internal::messageCapacity();
 }
-
-void recordForwarderMessageTruncation() { forwarder().noteTruncation(); }
-
+void recordForwarderMessageTruncation() {
+  forwarder_internal::noteTruncation();
+}
 // Public API: return forwarding counters snapshot.
-ForwarderStats getForwarderStats() { return forwarder().stats(); }
+ForwarderStats getForwarderStats() { return forwarder_internal::stats(); }
 
-// Public API: direct trace sink for already-formatted messages.
+// Public API: direct sinks for already-formatted messages.
 void trace(const std::string_view msg) {
   emit(LogLevel::trace, msg, "<unknown>", 0);
 }
-// Public API: direct info sink for already-formatted messages.
 void info(const std::string_view msg) {
   emit(LogLevel::info, msg, "<unknown>", 0);
 }
-// Public API: direct debug sink for already-formatted messages.
 void debug(const std::string_view msg) {
   emit(LogLevel::debug, msg, "<unknown>", 0);
 }
-// Public API: direct warn sink for already-formatted messages.
 void warn(const std::string_view msg) {
   emit(LogLevel::warn, msg, "<unknown>", 0);
 }
-// Public API: direct error sink for already-formatted messages.
 void error(const std::string_view msg) {
   emit(LogLevel::error, msg, "<unknown>", 0);
 }
 
-// Strip directory prefix to keep compact source location output.
+// Strip directory prefix to keep compact source metadata output.
 std::string pathToFileName(const std::string_view fullFilePath) {
   const auto pos = fullFilePath.find_last_of("/\\");
   if (pos == std::string_view::npos)
@@ -573,14 +269,14 @@ void setLogLevel(const LogLevel level) {
 // Return current runtime logging threshold.
 LogLevel getLogLevel() { return gLogLevel.load(std::memory_order_relaxed); }
 
-// Flush primary sink and wait for forwarding queue to drain.
+// Flush primary sinks and wait for forwarded queue drain.
 void flushLogs() {
   std::fflush(stdout);
   std::fflush(stderr);
-  forwarder().flush();
+  forwarder_internal::flush();
 }
 
-// Entry point used by templated call sites after message formatting.
+// Entry points used by templated header helpers after formatting.
 void logMessageFormatted(LogLevel logLevel, std::string formattedMessage,
                          const char *fileName, int lineNo) {
   emit(logLevel, formattedMessage, fileName, lineNo);
@@ -591,7 +287,6 @@ void logMessageView(LogLevel logLevel, std::string_view formattedMessage,
   emit(logLevel, formattedMessage, fileName, lineNo);
 }
 
-// Entry point for timestamped log helper after message formatting.
 void logWithTimestampFormatted(std::string formattedMessage,
                                const char *fileName, int lineNo) {
   emit(LogLevel::info, formattedMessage, fileName, lineNo);
